removed unused files and code
