+++ examples/example_utils/omezarr_metadata.py
